+++ db/migrate/20190611095055_create_patient_histories.rb
@@ -1,7 +1,7 @@
 class CreatePatientHistories < ActiveRecord::Migration[5.2]
   def change
     create_table    :patient_histories, :primary_key => :history_id do |t|
-    	t.integer       :encounter_id
+    	t.bigint       :encounter_id
     	t.integer       :concept_id
     	t.integer       :value_coded
     	t.boolean     :voided, null: false, default: 0
