+++ db/migrate/20190611095144_create_family_plannings.rb
@@ -1,7 +1,7 @@
 class CreateFamilyPlannings < ActiveRecord::Migration[5.2]
   def change
     create_table   :family_plannings , :primary_key => :family_planning_id do |t|
-    	t.integer      :encounter_id
+    	t.bigint      :encounter_id
     	t.integer      :concept_id
     	t.integer      :value_coded
     	t.boolean    :voided, null: false, default: 0
